Hi,

I've replaced the benchmark logic with `benchmark-ips` gem that is IMO much better for comparing such fast methods. Also I replaced the times block loop with the (by `benchmark-ips`) recommended `while` that has lower overhead.

Results on my machine (Ubuntu / ruby 2.2.0)
```
Calculating -------------------------------------
       select (enum)     1.005k i/100ms
       filter (enum)   315.000  i/100ms
      select (array)     1.218k i/100ms
      filter (array)   308.000  i/100ms
-------------------------------------------------
       select (enum)     10.101k (± 2.0%) i/s -     51.255k
       filter (enum)      3.185k (± 3.1%) i/s -     16.065k
      select (array)     12.824k (± 7.9%) i/s -     64.554k
      filter (array)      3.314k (± 2.0%) i/s -     16.632k

Comparison:
      select (array):    12824.3 i/s
       select (enum):    10101.5 i/s - 1.27x slower
      filter (array):     3313.8 i/s - 3.87x slower
       filter (enum):     3184.7 i/s - 4.03x slower

Calculating -------------------------------------
          map (enum)     1.002k i/100ms
       T. map (enum)   287.000  i/100ms
         map (array)     1.431k i/100ms
      T. map (array)   278.000  i/100ms
-------------------------------------------------
          map (enum)     10.249k (± 3.2%) i/s -     52.104k
       T. map (enum)      2.757k (± 2.7%) i/s -     13.776k
         map (array)     13.956k (± 3.7%) i/s -     70.119k
      T. map (array)      2.817k (± 3.1%) i/s -     14.178k

Comparison:
         map (array):    13956.1 i/s
          map (enum):    10248.6 i/s - 1.36x slower
      T. map (array):     2817.3 i/s - 4.95x slower
       T. map (enum):     2756.9 i/s - 5.06x slower

Calculating -------------------------------------
 map + select (enum)   550.000  i/100ms
 map + filter (enum)   219.000  i/100ms
map + select (array)   659.000  i/100ms
map + filter (array)   222.000  i/100ms
-------------------------------------------------
 map + select (enum)      5.647k (± 1.8%) i/s -     28.600k
 map + filter (enum)      2.208k (± 1.6%) i/s -     11.169k
map + select (array)      6.673k (± 3.4%) i/s -     33.609k
map + filter (array)      2.272k (± 2.4%) i/s -     11.544k

Comparison:
map + select (array):     6673.4 i/s
 map + select (enum):     5646.8 i/s - 1.18x slower
map + filter (array):     2271.9 i/s - 2.94x slower
 map + filter (enum):     2208.1 i/s - 3.02x slower
```