@dchelimsky 2015-02-07
Thanks for the contribution @jsuchal, but this we don't accept PRs for this project (https://github.com/cognitect-labs/transducers-ruby#contributing). I like the comparison output, so I'll take a look at benchmark-ips (and using while loops), though I'll probably group things slightly differently.