@jsuchal 2015-02-07
@dchelimsky no worries. I wanted to see the real difference myself. It's "throwaway code". Keep up the good work.